backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
groq==0.4.2
pandas==2.2.0
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic==2.5.0
openpyxl==3.1.2
requests==2.31.0
